readme update and small fixes

+++ cleanse.js
@@ -4,7 +4,7 @@ const cleanse = (elements) =>
   elements.map((element) => {
     return {
       url: element.url,
-      value: parseInt(element.element.replace("$", "")).toFixed(2),
+      value: parseFloat(element.element.replace("$", "")).toFixed(2),
     };
   });
